+++ packages/expo-router/_ctx.js
@@ -1,1 +1,6 @@
-export const ctx = require.context(process.env.EXPO_ROUTER_APP_ROOT, true, /.*/);
+export const ctx = require.context(
+  process.env.EXPO_ROUTER_APP_ROOT,
+  true,
+  // Ignore root `./+html.js` and API route files `./generate+api.tsx`.
+  /^(?:\.\/)(?!(?:(?:(?:.*\+api)|(?:\+html)))\.[tj]sx?$).*\.[tj]sx?$/
+);